Adds pagination component
Introduces a reusable pagination component for improved data navigation.

The component allows users to navigate through data sets, displaying the current range of items and the total number of items.
It uses a DTO to manage pagination state and emits events when the page changes.
CSS is included to set up the component and make it horizontally scrollable if needed.