Renders plan de estudios in a graph
Creates a component to visualize the plan de estudios as an interactive graph.

Fetches data from several endpoints (Carreras, Semestres, Asignaturas, Prerrequisitos)
to construct the graph data.

The graph displays asignaturas, their prerequisites, and the overall structure of the plan de estudios.
A loading spinner is shown while the data is being fetched.

The diagram is rendered using the goJS library.