Adds related subjects visualization
Implements a feature to display related subjects for a given subject.

This change introduces a new button in the subject list table
that, when clicked, displays a graph of associated subjects.

It also adjusts the table's maximum height to accommodate the new graph,
and adds a button to remove the graph from the view.